DataInitializer, .gitignore file and some web init files added
